Fix crash when using 2.3.1 in UICollectionViews or similar
In certain cases (collection views just being one of them), view
initialization takes place within a `layoutSubviews`  call.

-[UIView layoutSubviews] does some evaluation of constraints and seems
to keep its own references to them. Therefore it is possible, that
constraints of the placeholderView are being used, after we're done
with view initialization.

As the `placeholderView` gets replaced it would previously be
deallocated immediately. As NSLayoutConstraints only keep weak
references on their items this lead to a dangling pointer on
the `firstItem` property.

This change asynchronously dispatches a block enclosing the
`placeholderView`. This defers the deallocation of the `placeholderView`
until the end of the current runloop cycle and allows the current
layoutSubviews call to go through before.